requirements.txt
Django==2.2
django-ckeditor==6.0.0
django-js-asset==1.2.2
Pillow==7.2.0
pkg-resources==0.0.0
pytz==2020.1
sqlparse==0.3.1
